Move logins sync-schema to separate file
